Make【サーバーサイド】詳細表示機能

+++ resources/views/posts/show.blade.php
@@ -15,30 +15,36 @@
 
                     <div class="alert alert-primary" role="alert">
                       <span>タイトル：</span>
-                      タイトル
+                      {{ $post->title}}
                     </div>
                     <div class="mb-3">
                         <img src="/storage/no-image.png" class="img-responsive d-block mx-auto" style="max-width: 100%; height: auto; width /***/:auto;">
                     </div>
                     
-                    <div class="border-bottom pb-2">内容</div>
+                    <div class="border-bottom pb-2">
+                    <span class="ml-3" >{{ $post->user->name}}：：</span>{{ $post->content}}</div>
                     <div class="border-bottom pt-2 pb-2">
                     
-                    <span class="ml-3">投稿日：：</span>00/00/00
+                    <span class="ml-3">投稿日：：</span>{{ $post->created_at}}
                     </div>
+
+                    {{-- 現在ログインしているユーザーの投稿であれば表示する --}}
+
+                    @auth
+                    　@if(($post->user_id) === (Auth::user()->id ))
                       <div class="d-flex">
                         <a href="">
-                          <br>
                           <input class="btn btn-primary" type="submit" value="変更する">
                         </a>
                         <div class="ml-3">
                           <form method="" action="" id="">
                             @csrf
-                            <br>
                             <a href="#" class="btn btn-danger" data-id="" onclick="deletePost(this); ">削除する</a>
                           </form>
                         </div>
                       </div>
+                      @endif
+                    @endauth
                 </div>
             </div>
         </div>
